[ROCm] Move HIP-only compat files out of source/ into hip/
Relocate the two HIP-only pieces of the AMD build -- the force-included
cuda_to_hip.h alias header and the hip_compat/ toolkit-header forwarding
shims -- from under source/ to a new top-level hip/ directory, so that
source/ holds only the CUDA-spelled engine and no HIP-specific files.

cuda_to_hip.h moves to hip/cuda_to_hip.h; the shim tree moves verbatim to
hip/hip_compat/. The shims previously reached the alias header via
<EngineKernels/cuda_to_hip.h> (resolved through source/ on the include
path); they now include it as <cuda_to_hip.h>, and CMake prepends the new
hip/ directory to the HIP include path so that resolves. The force-include
path and the hip_compat include directory are repointed to hip/.

Every edit is inside an if(USE_HIP) branch. The default CUDA build never
referenced these files at their old location and does not reference them at
the new one, so it is byte-identical in behavior: it resolves the real CUDA
toolkit headers and compiles the device kernels unchanged.

This work was done with the assistance of Claude, an AI coding assistant.

Test Plan:

USE_HIP=ON build and full GPU test suite on an AMD Instinct MI250X (gfx90a),
ROCm 7.2.1:

```
cmake -S . -B build -G Ninja \
  -DCMAKE_TOOLCHAIN_FILE=external/vcpkg/scripts/buildsystems/vcpkg.cmake \
  -DCMAKE_BUILD_TYPE=Release -DUSE_HIP=ON -DCMAKE_HIP_ARCHITECTURES=gfx90a \
  -DCMAKE_HIP_COMPILER=/opt/rocm/llvm/bin/clang++ -DCMAKE_PREFIX_PATH=/opt/rocm
cmake --build build -j 16
HIP_VISIBLE_DEVICES=0 ./build/EngineInterfaceTests --gtest_filter=-GeometryTests.*
HIP_VISIBLE_DEVICES=0 ./build/PersisterTests
HIP_VISIBLE_DEVICES=0 ./build/EngineTests \
  --gtest_filter=-GeometryTests.*:-NeuronPerformanceTests.*
```

EngineInterfaceTests 159/159, PersisterTests 64/64. EngineTests: 3042 ran,
3037 passed, 3 skipped, 2 failed. The 2 failures are pre-existing
platform-independent non-defects also present before this change:
CommunicatorTests.sender_signalPriority_lowerNumTimesSentWins (an upstream
last-writer-wins race whose winner depends on block execution order) and
DataTransferTests.multipleCells_genome_multipleGenes_multipleNodes (an
exact-float equality check tripped by a ~1 ULP rounding difference on a pure
set/get round-trip). Default CUDA path re-checked with nvcc (USE_HIP=OFF):
the EngineKernels device-code library compiles and device-links unchanged.
