+++ hip/hip_compat/cooperative_groups/reduce.h
