+++ hip/hip_compat/cuda/helper_string.h
