+++ hip/hip_compat/cuda_gl_interop.h
@@ -1,3 +1,3 @@
 #pragma once
 #include <hip/hip_gl_interop.h>
-#include <EngineKernels/cuda_to_hip.h>
+#include <cuda_to_hip.h>
