+++ hip/hip_compat/cuda_runtime_api.h
@@ -1,0 +1,2 @@
+#pragma once
+#include <cuda_to_hip.h>
