+++ hip/hip_compat/device_launch_parameters.h
@@ -1,3 +1,3 @@
 #pragma once
 // HIP defines threadIdx/blockIdx/blockDim/gridDim in <hip/hip_runtime.h>.
-#include <EngineKernels/cuda_to_hip.h>
+#include <cuda_to_hip.h>
