+++ hip/hip_compat/sm_60_atomic_functions.h
@@ -1,3 +1,3 @@
 #pragma once
 // CUDA's sm_60 atomics (atomicAdd on double, etc.) are part of the HIP runtime.
-#include <EngineKernels/cuda_to_hip.h>
+#include <cuda_to_hip.h>
