+++ hip/hip_compat/vector_types.h
@@ -1,3 +1,3 @@
 #pragma once
 // float2/int2/uchar4/... are provided by <hip/hip_runtime.h>.
-#include <EngineKernels/cuda_to_hip.h>
+#include <cuda_to_hip.h>
