--- source/hip_compat/cuda_runtime_api.h
@@ -1,2 +1,0 @@
-#pragma once
-#include <EngineKernels/cuda_to_hip.h>
